:tractor: Continue refactoring by moving the states code
